+++ src/base/b-list/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.??.?? (2023-??-??)
+## v3.52.0 (2023-07-19)
 
 #### :rocket: New Feature